app/requirements.txt
aiofiles==0.7.0
asgiref==3.3.4
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.65.2
flake8==3.9.2
greenlet==1.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.20.3
passlib==1.7.4
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.18
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
